requirements.txt
Django==4.2.2
django_allauth==0.54.0
docx==0.2.4
langchain==0.0.237
openai==0.27.8
python_docx==0.8.11
django-crispy-forms
django-bootstrap4
crispy-bootstrap4
pypdf
pycryptodome
openai
docx
tiktoken
faiss-cpu
django-inlinecss
psycopg2
pymupdf
django-recaptcha
django-simple-captcha
captcha